% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SSLASSO.R
\name{plot.SSLASSO}
\alias{plot.SSLASSO}
\title{Plot coefficients from a "SSLASSO" object}
\usage{
\method{plot}{SSLASSO}(x, ...)
}
\arguments{
\item{x}{An object of class "SSLASSO", usually, a result of a call to \code{SSLASSO}.}

\item{...}{Other arguments to be passed to \code{plot}.}
}
\value{
No return value, called for side effects (produces a plot of the coefficient paths).
}
\description{
Produces a coefficient profile plot of the SSLASSO solution path.
}
